requirements.txt
djangorestframework
djangorestframework-simplejwt
psycopg2-binary 
redis
django-redis
python-dotenv
